02_back_end/requirements.txt
fastapi[all]
uvicorn
mysql-connector-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings